Better scaling.
Don't scale cells that are smaller then the width of the table divided by the amount of columns.

Otherwise I get ridiculous cut offs if one cell is long and the other is way shorter (for example a long descriptive text in one column, in the other column a short value).